Improvements with Providers

+++ lib/src/providers/core/formatter.provider.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'package:intl/intl.dart';
 
+import '../interfaces/interfaces.dart';
 import 'instrument.provider.dart';
-import 'interfaces/interfaces.dart';
 
 const _kEmptyString = '';
 
@@ -19,7 +19,7 @@ class FormatterProvider implements AbstractFormatterProvider {
     return instrumentProvider.metadata(code).take(1).map((instrumentMetadata) {
       if (instrumentMetadata != null) {
         final format = instrumentMetadata.format;
-        final round = format.round as int;
+        final round = format.round;
         final formatter =
             NumberFormat.simpleCurrency(locale: locale, name: code);
 
@@ -43,7 +43,7 @@ class FormatterProvider implements AbstractFormatterProvider {
     return instrumentProvider.metadata(code).take(1).map((instrumentMetadata) {
       if (instrumentMetadata != null) {
         final pip = instrumentMetadata.pip;
-        final round = pip.round as int;
+        final round = pip.round;
         final formatter =
             NumberFormat.simpleCurrency(locale: locale, name: code);
